+++ src/be_atom_memo_tuple5_/index.d.ts
@@ -34,7 +34,7 @@ export declare function be_atom_memo_tuple5_<
 	ctx_T extends Ctx = Ctx
 >(
 	id:string|null|undefined,
-	atom__new:((ctx:MapCtx)=>A),
+	atom__new:(ctx:MapCtx)=>A,
 	be__params?:be__params_T
 ):be_atom_memo_tuple5_T<A, ctx_T>
 export type be_atom_memo_tuple5_T<