+++ src/val__be_atom_memo_tuple5_/index.d.ts
@@ -1,5 +1,6 @@
 import type { Autosubscribe, WritableAtom_ } from '@ctx-core/nanostores'
 import type { Be, be__params_T, Ctx, MapCtx } from '@ctx-core/object'
+import { BoxAutosubscribe, UnboxAutosubscribe } from 'nanostores'
 export declare function val__be_atom_memo_tuple5_<
 	V,
 	ctx_T extends Ctx = Ctx
@@ -25,17 +26,19 @@ export declare function val__be_atom_memo_tuple5_<
 	V,
 	ctx_T extends Ctx = Ctx
 >(
-	val__new:(ctx:MapCtx)=>V|PromiseLike<V>,
+	val__new:(ctx:MapCtx)=>
+		BoxAutosubscribe<V>|PromiseLike<BoxAutosubscribe<V>>,
 	be__params?:be__params_T
-):val__be_atom_memo_tuple5_T<Awaited<V>, ctx_T>
+):val__be_atom_memo_tuple5_T<UnboxAutosubscribe<Awaited<V>>, ctx_T>
 export declare function val__be_atom_memo_tuple5_<
 	V,
 	ctx_T extends Ctx = Ctx
 >(
 	id:string|null|undefined,
-	val__new:((ctx:MapCtx)=>V|PromiseLike<V>),
+	val__new:(ctx:MapCtx)=>
+		BoxAutosubscribe<V>|PromiseLike<BoxAutosubscribe<V>>,
 	be__params?:be__params_T
-):val__be_atom_memo_tuple5_T<Awaited<V>, ctx_T>
+):val__be_atom_memo_tuple5_T<UnboxAutosubscribe<Awaited<V>>, ctx_T>
 export type val__be_atom_memo_tuple5_T<
 	V,
 	ctx_T extends Ctx = Ctx